more abstraction, fix type switch
